feature/shared-license-stage1

* fix: remove notification garbage clean api

* chore: comments

* fix: migration

* chore: update doc

* Chore: documents

* Fix: one active license package per type

* Fix: items[].price

* Fix: items[].price

* chore: update test models

* Merge branch 'dev2' into feature/shared-license-stage1

* Merged in fix/subscription-price-and-item-price (pull request #238)

fix/subscription-price-and-item-price

* wip: fix subscripition items and item price/tax/amount


Approved-by: Shengtian Liao

* Merged in function/purchase-licence-package (pull request #236)

function/purchase-licence-package

* update api

* wip: migration

* update: LicensePackage.cpp

* wip: Subscription.fillLicensePackage()

* wip: license package count => quantity

* wip: subscirption::buildItems() & fillLicensePackage()

* wip: subscription attribute options

* wip: remove subscription.fillLicensePackage

* wip: subscription.items and next invoice

* wip: invoice.items & license_package_info

* wip: test cases fix

* wip: test helper

* test: LicensePackage.refreshInfo()

* createSubscription

* Fix: invoice controller

* wip: notifications

* wip: refactor SubscriptionPlan

* wip: upgrade packages

* wip: refactor EmailHelper.php and update email blade

* update setting.json

* wip: migrate data

* wip: rename migration name

* Merged in function/rename-user-license-count (pull request #237)

wip: user.license_count => seat_count

* wip: user.license_count => seat_count


Approved-by: Shengtian Liao

* fix: bug to create license sharing

* Merge branch 'function/purchase-licence-package' of bitbucket.org:futurecorporation/siser-backend into function/purchase-licence-package

* wip: /license-sharings and /license/-sharings/{id}

* wip: design.md

* wip: add license_count back

* wip: test models

* Merge branch 'feature/shared-license-stage1' into function/purchase-licence-package


Approved-by: Shengtian Liao

* Refactor: change license_package_info.count => quantity

* Merge branch 'dev2' into feature/shared-license-stage1

* Merged in function/invitation-management (pull request #232)

function/invitation-management

* Merge branch 'feature/shared-license-stage1' into function/invitation-management

* wip: update manager dr

* wip: license sharing invitation account update & delete test

* wip: license sharing invitation create test

* test: license sharing invitation account get

* licese sharing invitation account list test

* wip: license sharing invitation to me update api test

* remove useless code

* wip: license sharing to me account list

* wip: refactor license sharing test helper

* wip: license sharing account get and list

* wip: license sharing account list test

* wip: license-sharing test and command

* wip: test cases

* wip: fix bugs

* wip: fix license sharing refactor

* wip: create LicenseSharingService.php

* wip: license sharing notification

* wip: notification test

* Rename TestController.php

* Merge branch 'feature/shared-license-stage1' into function/invitation-management

* wip: fix LicenseSharing::updateFromSubscripiton()

* Fix & mockup

* Fix: license sharing

* wip: fix license-shareing - test

* wip: LicenseSharingController.php

* Fix: license sharing list filter

* Fix: LicenseSharingInvitationController.php validations

* wip: api endpoint

* Merge branch 'feature/shared-license-stage1' into function/invitation-management

* wip: license sharing invitation control place holder

* wip: licenseSharing & Invitation models and migrations

* wip: refactor license sharing invitation

* wip: add owner/guest name & email

* wip: LicenseSharing.php

* Merge branch 'feature/shared-license-stage1' into function/invitation-management

* wip: invitation manage draft

* wip: update generated models (invoice, license-package, subscription)


Approved-by: Shengtian Liao

* Merge branch 'dev2' into feature/shared-license-stage1

* Merge branch 'dev2' into feature/shared-license-stage1

* Merged in change/license-package-management (pull request #231)

change/license-package-management

* func: migration and model for LicensePackage

* wip: update LicensePackage and add LicensePackageInfo class

* wip: remove subscription's extra items

* wip: add controller code

* wip: license package test cases

* wip: update test models

* Merge branch 'feature/shared-license-stage1' into function/license-package-management

* wip: fix license-package model & migration

* wip: update license-package-controller & test

* wip: fix license package validation

* Merge branch 'feature/shared-license-stage1' into change/license-package-management


Approved-by: Shengtian Liao

* Merge branch 'dev2' into feature/shared-license-stage1

* revert: remove extra items for now.

* Merged in function/license-package-management (pull request #230)

function/license-package-management

* wip: update test models

* wip: license package test cases

* wip: add controller code

* wip: remove subscription's extra items

* wip: update LicensePackage and add LicensePackageInfo class

* func: migration and model for LicensePackage


Approved-by: Shengtian Liao

* Add invoice.isCancellable()

Approved-by: Shengtian Liao

+++ bin/rebuild-data.sh
@@ -1,9 +1,16 @@
 #!/bin/bash
 
 set -e
-set -x
+# set -x
 
 BASEDIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" &>/dev/null && pwd)"
+
+# only run in local environment
+APP_ENV=$(php $BASEDIR/../artisan env | sed -n 's/.*\[\(.*\)\].*/\1/p')
+if [ "$APP_ENV" != "local" ]; then
+  echo "Script stopped because APP_ENV is not local"
+  exit 1
+fi
 
 php $BASEDIR/../artisan migrate:fresh
